backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.14,<4.0
boto3>=1.26,<2.0
python-dotenv>=1.0,<2.0
dj-database-url>=1.0,<2.0